Rename favicon and manifest to bybass CloudFront caching
Since all assets except index.html are cached for 30 days which is fine
for js and css since their names are dynamically created BUT not fine
for these two since their names are always the same hence they are
served from the cache which expires during the next 30 days. And I don't
want to wait that long so this is why I am doing this.